Simplify development process
- remove building configs
- bundle source files
- add demo page from dev
