+++ .gitignore
@@ -1,5 +1,3 @@
-dev/
-dist/
 .idea/
 node_modules/
 .vscode/
