--- src/autoinit.js
@@ -1,8 +1,0 @@
-import initializeShareon from './shareon';
-import './style.css';
-
-window.onload = () => {
-  initializeShareon();
-};
-
-export default initializeShareon;
